Rework selecting device on connect.
Only try to connect to the device if the selection is not the first,
"Select MIDI device" entry. Otherwise we'd end up in trying to connect
to that in various situations, making the connection only work in some
cases.

Also, since we now only connect if we have both an input and an output
port connected always reconnect in the callback.