Allow slider type to also have a value list.
Support handling a parameter that shown as a range to also be
represented as option, i.e. additionally contain a value list.
Allows handing the named tempi values the same way as the tempo slider
is handled, and removes the special case needed before.

Currently only implemented for the range -> option case. Requires the
option tag to use the new "for" attribute to point to the actual value,
and updating the option value needs to be done in the value callback.