Make init status tristate and track it.
On connection we read several memory areas to initially update the view.
Track if we ever got responses for those and change the status to a
yellow icon unless we have all values.

This makes it clear if we have a MIDI connection set up, or both a
connection and all values initialized.