Support connection via BLE.
If the browser supports BLE add an extra "Bluetooth" item to the device
dropdown. Handle the BLE connection and its additional protocol bytes.

Initializing the connection via BLE is a bit slow, typically the first
requests aren't responsed by the instrument. Add a retry mechanism,
which should work better than simply putting in a delay, and also avoids
problems if the instrument only partially responds to the initial
requests.